requirements.txt
tensorflow==2.19.0
keras==3.10.0
numpy==2.0.2
streamlit
Pillow
opencv-python-headless
matplotlib
